+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - YYYY-MM-DD
 
 ### Added
-* Italian support
+* Bahasa Indonesia, Nederands, and Vietnamese support
+
 
 ### Changed
 
@@ -20,6 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Fixed
+
+
+## [v0.1.4] - 2024-10-09
+
+### Added
+* Bahasa Indonesia support
+* Nederands support
+* Vietnamese support
 
 
 ## [v0.1.3] - 2024-10-06